# openapi.yaml
openapi: 3.1.0
info:
  title: Local Translation Service
  version: 0.1.0
  description: >-
    Local pluggable machine translation HTTP API. All HTTP bodies are limited by
    MAX_REQUEST_BODY_BYTES (default 16000000 bytes) before JSON parsing; exceeding
    this limit returns 413 request_body_too_large.
servers:
  - url: http://localhost:8000
paths:
  /v1/annotate:
    post:
      summary: Annotate content
      operationId: annotateContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotateRequest'
            examples:
              url:
                value: {url: 'https://example.org/article'}
              html:
                value: {html: '<main><p>Article content.</p></main>'}
              text:
                value: {text: 'Article content.'}
              localQwen:
                value: {text: 'Article content.', annotator: qwen_local}
      responses:
        '200':
          description: Annotation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotateResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '413':
          $ref: '#/components/responses/Error413'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '502':
          $ref: '#/components/responses/Error502'
        '503':
          $ref: '#/components/responses/Error503'
        '504':
          $ref: '#/components/responses/Error504'
  /v1/annotators:
    get:
      summary: List registered annotators
      operationId: listAnnotators
      responses:
        '413':
          $ref: '#/components/responses/Error413'
        '200':
          description: Annotator registry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotatorListResponse'
  /v1/translate:
    post:
      summary: Translate text
      operationId: translateText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateRequest'
            examples:
              minimal:
                value:
                  text: The spacecraft entered orbit.
                  target_language: ru
              explicitTranslator:
                value:
                  text: The spacecraft entered orbit.
                  target_language: ru
                  source_language: en
                  translator: argos
                  translator_params: {}
              marianTranslator:
                value:
                  text: A collection of media links about air quality research.
                  target_language: ru
                  source_language: en
                  translator: marian
                  translator_params: {}
      responses:
        '200':
          description: Translation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '404':
          $ref: '#/components/responses/Error404'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
        '504':
          $ref: '#/components/responses/Error504'
  /v1/translators:
    get:
      summary: List registered translators
      operationId: listTranslators
      responses:
        '413':
          $ref: '#/components/responses/Error413'
        '200':
          description: Translator registry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslatorListResponse'
  /health/live:
    get:
      summary: Liveness probe
      operationId: healthLive
      responses:
        '413':
          $ref: '#/components/responses/Error413'
        '200':
          description: Process is alive
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
  /health/ready:
    get:
      summary: Readiness probe
      operationId: healthReady
      responses:
        '413':
          $ref: '#/components/responses/Error413'
        '200':
          description: Default translator is ready
          content:
            application/json:
              schema:
                type: object
                required: [status, default_translator]
                properties:
                  status:
                    type: string
                    const: ready
                  default_translator:
                    type: string
        '503':
          $ref: '#/components/responses/Error503'
components:
  schemas:
    AnnotateRequest:
      type: object
      additionalProperties: false
      description: Exactly one of url, html, or text must be provided.
      oneOf:
        - required: [url]
        - required: [html]
        - required: [text]
      properties:
        url:
          type: string
          description: Public HTTP(S) URL.
        html:
          type: string
          description: HTML cleaned before annotation.
        text:
          type: string
          description: Text passed unchanged to the annotator.
        annotator:
          type: string
          description: Uses DEFAULT_ANNOTATOR when omitted. Built-in values are iishko and qwen_local.
        annotator_params:
          type: object
          additionalProperties: true
          default: {}
    AnnotateResponse:
      type: object
      additionalProperties: false
      required: [annotation, language, annotator, metadata]
      properties:
        annotation:
          type: string
        language:
          type: string
          const: ru
        annotator:
          type: string
        metadata:
          type: object
          additionalProperties: false
          required: [duration_ms, truncated]
          properties:
            duration_ms: {type: integer, minimum: 0}
            truncated: {type: boolean}
    AnnotatorDescriptor:
      type: object
      additionalProperties: false
      required: [name, ready]
      properties:
        name: {type: string}
        ready: {type: boolean}
    AnnotatorListResponse:
      type: object
      additionalProperties: false
      required: [annotators, default_annotator]
      properties:
        annotators:
          type: array
          items:
            $ref: '#/components/schemas/AnnotatorDescriptor'
        default_annotator: {type: string}
    LanguageCode:
      type: string
      pattern: '^[a-z]{2,3}(-[A-Za-z0-9]+)*$'
      examples: [en, ru]
    TranslateRequest:
      type: object
      additionalProperties: false
      required:
        - text
        - target_language
      properties:
        text:
          type: string
          minLength: 1
          description: >-
            Source text. Runtime maximum is configured by MAX_TEXT_LENGTH. Marian
            additionally limits input to the tokenizer model_max_length, capped at
            1024 tokens including special tokens; exceeding it returns 413 text_too_large.
            Incomplete Marian generation returns 500 translation_failed.
        target_language:
          $ref: '#/components/schemas/LanguageCode'
        source_language:
          allOf:
            - $ref: '#/components/schemas/LanguageCode'
          description: Optional. Uses DEFAULT_SOURCE_LANGUAGE when omitted.
        translator:
          type: string
          minLength: 1
          description: Optional provider name. Uses DEFAULT_TRANSLATOR when omitted.
        translator_params:
          type: object
          additionalProperties: true
          default: {}
          description: Provider-specific parameters validated by the selected provider.
    TranslateResponse:
      type: object
      additionalProperties: false
      required:
        - translation
        - source_language
        - target_language
        - translator
        - metadata
      properties:
        translation:
          type: string
        source_language:
          $ref: '#/components/schemas/LanguageCode'
        target_language:
          $ref: '#/components/schemas/LanguageCode'
        translator:
          type: string
        metadata:
          type: object
          required: [duration_ms]
          additionalProperties: true
          properties:
            duration_ms:
              type: integer
              minimum: 0
    LanguagePair:
      type: object
      additionalProperties: false
      required: [source, target]
      properties:
        source:
          $ref: '#/components/schemas/LanguageCode'
        target:
          $ref: '#/components/schemas/LanguageCode'
    TranslatorDescriptor:
      type: object
      additionalProperties: false
      required:
        - name
        - ready
        - supported_language_pairs
      properties:
        name:
          type: string
        ready:
          type: boolean
        supported_language_pairs:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/LanguagePair'
            - type: 'null'
    TranslatorListResponse:
      type: object
      additionalProperties: false
      required: [translators, default_translator]
      properties:
        translators:
          type: array
          items:
            $ref: '#/components/schemas/TranslatorDescriptor'
        default_translator:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
  responses:
    Error400:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error403:
      description: URL is not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error404:
      description: Unknown provider
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error413:
      description: Content, Marian tokenized input, or HTTP body exceeds its limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error415:
      description: Unsupported content type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error422:
      description: Invalid parameters or unsupported content
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error500:
      description: Translation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error502:
      description: Fetch or annotation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error503:
      description: Provider unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error504:
      description: Operation timeout
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'